get grudges route test added
