+++ package.json
@@ -27,6 +27,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
+    "mtest": "mocha",
     "eject": "react-scripts eject",
     "dev": "nodemon server/server.js"
   }
